- "Sprite" is now same as English "Sprite" instead of "Duszek" (Little Ghost).
Polish equivalent never took root in gamedev (while it tried since early '90s). Some documentations like Defold, SFML are using not translated "sprite" in Polish translation. Also, Polish gamedev forums, communities and books are using "sprite" in 99% cases, or Ponglish version - "sprajt". I suspect it's because exact translation to "ghost" is associated with rather translucent thing, which is not true for many sprites, and other equivalents to sprite like "pixie, leprechaun, dwarf, gnome" are even worse. I didn't want to use "picture/image/icon/painting/illustration" which for sure was closer description of it in Polish, as that could be confused with backgrounds or images set in Game Options, and also tried to avoid archaic words.